Added link to screen values, buttons & mathEngine
Used template to display the values of the lower and upper display properties.

Added the directives to host the fourth and fith row of buttons.

Also created mathEngine service to handle all calculations when triggered by
the orchestrator service. This makes use of the mathjs library and types in
@types. The library is added to the global namespace and so there is no need
to import.