Resolved mathjs import & refactored ScreenComponent
Resolved mathjs import:
Manually (not the best way) linked all required modules within and
outside mathjs (but still in node_modules). This was very painful,
it can be refactored :). This was necessary because the mathjs modules
have index.js files as the main entry point within them, as opposed to
moduleName.js (Rxjs has the latter).

Refactored Screen Component:
Refactored the OrchestratorService so that the ScreenComponent is
responsible for deciding what to do with the value that's just been
pushed. I left the accommodation for this in the OrchestratorService
just incase.