Added CE Operation & Number Formatting
CE Operation:
The purpose of the CE operation is to clear the last input, be it a
digit or operator. This is necessary to allow users correct mistakes.

Number Formatting:
The output from any calculation has been limited to 9 decimal places.
This has been implemented in the mathEngine.service and was done to
prevent numbers from overflowing out of the UpperScreen.

Minor changes:
1) Change to Calculator width