Integrated Karma Test Runner & Jasmine Tests
Karma-test.shim & Karma.conf.js were added to integrate tests.
The locations of the mathjs libraries needed to be added to the
karma config file to ensure they were added to the server the test-
runner starts up.

Minor changes were also made to remove dead code in the mathEngineService
& OrchestratorService. Added test for the mathEngine and a dummy test for
the CalcButtonComponent.