Integrated Webpack for Dev Start & Prod Build
Dev Start:
This was done as per the Angular Webpack Introduction on angular.io.
Various webpack loaders and dependencies were installed to help with inlining
templates. Import statements were all moved into components. The index.html no
longer loads css, the main AppComponent imports this. This was done to take
advantage of Webpack's bundling, uglification & minification process.
The awesome-typescript-loader that transpiles the ts files needs a config,
but if the config is not in the root directory, the location of the "types"
declarations need to be in the tsconfig.json. Without this, the code will have
errors on build and so will fail. This is why the typeRoots property is
filled and why the options are added to the loader config in webpack.common.

Build:
Same as above. It emits files to folder dist.