Fixed getting "undefined" if pressing '=' first
In previous releases. pressing "=" first on the calculator would give an
"undefined". By first, this means not pressing any numbers or anything, so the
working value store is basically empty.

Minor changes:
Fixed the sizes of the upper and lower screens so that when empty they don't
decompose to a line.